requirements.txt
flask
flask-login
flask-wtf
coloredlogs
python-dotenv
pymongo

